Refactored the entire IDA UI to be more readable
