+++ plugins/ida_binsync/ida_binsync/ui/binsync.png
